+++ lib/iruby.rb
@@ -7,7 +7,6 @@ require 'set'
 require 'iruby/version'
 require 'iruby/kernel'
 require 'iruby/backend'
-require 'iruby/session'
 require 'iruby/ostream'
 require 'iruby/formatter'
 require 'iruby/utils'
@@ -15,8 +14,11 @@ require 'iruby/display'
 require 'iruby/comm'
 
 begin
-  require 'rbczmq'
-rescue LoadError => e
-  require 'ffi-rzmq'
-  require 'iruby/session_ffi'
+  require 'iruby/session/rbczmq'
+rescue LoadeError => e
+  begin
+    require 'iruby/session/ffi_rzmq'
+  rescue LoadeError => e
+    STDERR.puts "You should install rbczmq or ffi_rzmq before running iruby notebook. See README."
+  end
 end
