+++ lib/iruby/session/ffi_rzmq.rb
