+++ lib/iruby/session/rbczmq.rb
